docker-compose/miner/requirements.txt
Flask
prophet
matplotlib
pmdarima
mysql-connector-python
numpy
pandas

